Fix: revise structure, keys and dates generate within modules
