+++ lib/datable.rb
@@ -1,0 +1,9 @@
+module Datable
+  def generate_date(date)
+    if date == nil
+      date = Date.today.strftime("%d%m%y") #"DDMMYY"
+    else
+      date
+    end
+  end
+end
