+++ spec/datable_spec.rb
@@ -1,0 +1,15 @@
+require './spec_helper'
+
+RSpec.describe Datable do
+  before(:each) do
+    @enigma = Enigma.new
+  end
+
+  it '#generate - returns given date' do
+    expect(@enigma.generate_date("040895")).to eq("040895")
+  end
+
+  it '#generate - returns todays date' do
+    expect(@enigma.generate_date(nil)).to eq(Date.today.strftime("%d%m%y"))
+  end
+end
