+++ spec/spec_helper.rb
@@ -2,6 +2,6 @@ require 'simplecov'
 SimpleCov.start
 
 require './lib/enigma'
-require './lib/key_gen'
-require './lib/date_gen'
+require './lib/keyable'
+require './lib/datable'
 require './lib/cipher'
